flask-backend/requirements.txt
sqlalchemy
flask
Flask-SQLAlchemy
sqlite3
flask-bcrypt